+++ CHANGELOG.md
@@ -4,6 +4,7 @@ v1.0.3 (2019-06-14)
 * Added CLITEMP internal temperature
 * Added cmd CLIMATE_ON to start conditioning the vehicle
 * Added cmd CLIMATE_OFF to stop conditioning the vehicle
+* Added CLIEMD Climate conditioning status
 * Fixed spelling of vehicle.
 * Fixed sunroof data parsing when the vehicle does not have a sunroof.